test: migrate WikiSelector tests in source
